pyproject.toml
[project]
name = "equity-alpha-pipeline"
version = "0.1.0"
description = "Factor-neutral long/short U.S. equity alpha research pipeline"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26,<2.1",
    "pandas>=2.2,<2.3",
    "scipy>=1.13,<1.15",
    "statsmodels>=0.14",
    "scikit-learn>=1.3",
    "cvxpy>=1.4",
    "plotly",
    "kaleido",
    "pyarrow>=12.0",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "pandas-datareader>=0.10",
]

[project.optional-dependencies]
ml = ["xgboost>=2.0"]
dev = ["pytest>=7.0", "pytest-cov", "ruff", "mypy"]
notebook = ["jupyterlab>=4.0", "ipywidgets"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
